reformat okapi namespace
