+++ okapi/.htaccess
@@ -2,4 +2,4 @@ Options -Indexes
 
 RewriteEngine on
 RewriteCond %{REQUEST_URI} !/okapi/static/
-RewriteRule ^(.*)$ OkapiScriptEntryPointController.php [L,QSA,E=no-gzip:1,E=dont-vary:1]
+RewriteRule ^(.*)$ okapi.php [L,QSA,E=no-gzip:1,E=dont-vary:1]
